requirements.txt
asgiref==3.3.4
Django==3.2
django-admin-interface==0.16.1
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-registration-redux==2.9
django-summernote==0.8.11.6
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
mysql-connector-python==8.0.23
Pillow==8.2.0
protobuf==3.15.8
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
